
\docType{data}
\name{camr}
\alias{camr}
\title{(Data) child mortality in China, cummulative by age and strata of MCHSS.}
\format{A data frame with 720 rows and 6 variables:
\describe{
  \item{agegp}{Region of China (east, middle, or west)}
  \item{reg}{Region of China (east, middle, or west)}
  \item{res}{rural or urban}
  \item{year}{Year of surveillance}
  \item{nmr}{Neonatal mortality (per live birth)}
  \item{u5mr}{Under five mortality (per live birth)}
    ...
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S2214109X16303345}
}
\usage{
u5mr
}
