\name{kcalc}
\alias{kcalc}

\title{Estimating specific life tables of fixed child ages}
\usage{
kcalc()
}
\description{
	To fit cause specific under 5 Q, when qxs are predicted directly
  To match cause specific neonatal mortality}
\examples{
kcalc()
}
\references{
Wilmoth J, Zureick S, Canudas-Romo V, Inoue M, Sawyer C. A flexible two-dimensional mortality model for use in indirect estimation. \emph{Population studies}. 2012 Mar 1;66(1):1-28.
}