
\docType{data}
\name{u5mr}
\alias{u5mr}
\title{(Data) Under five and neonatal mortality in China.}
\format{A data frame with 120 rows and 5 variables:
\describe{
  \item{reg}{Region of China (east, middle, or west)}
  \item{res}{rural or urban}
  \item{year}{Year of surveillance}
  \item{nmr}{Neonatal mortality (per live birth)}
  \item{u5mr}{Under five mortality (per live birth)}
    ...
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S2214109X16303345}
}
\usage{
u5mr
}
