
\docType{data}
\name{u5mr}
\alias{u5mr}
\title{(Data) Under five and neonatal mortality in China.}
\format{A data frame with 120 rows and 5 variables:
\describe{
  \item{reg}{Region of China (east, middle, or west)}
  \item{res}{rural or urban}
  \item{year}{Year of surveillance}
  \item{nmr}{Neonatal mortality (per live birth)}
  \item{u5mr}{Under five mortality (per live birth)}
    ...
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S2214109X16303345}
}
\usage{
u5mr
}
\references{
He C, Liu L, Chu Y, Perin J, Dai L, Li X, Miao L, Kang L, Li Q, Scherpbier R, Guo S. National and subnational all-cause and cause-specific child mortality in China, 1996–2015: a systematic analysis with implications for the Sustainable Development Goals. The Lancet Global Health. 2017 Feb 1;5(2):e186-97.
}